requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai>=1.54.0
httpx>=0.27.0,<0.28.0
pydantic==2.9.2
websockets>=12.0
huggingface-hub==0.27.0